Fixed freezing when zoomed a lot with the doubles. Now when it reaches the end of its resolution, it starts 'zooming in' like it did with floats

+++ src/bkow/meow/processing/linear/predicate/error/visualization/sketch/ColorScreen.java
@@ -72,7 +72,7 @@ public class ColorScreen {
         for (int j = 0; j < GlobalVar.gs().getHeight(); j++){
           c = m_screen[i][j];
           BoxPoint p = GlobalVar.ps().pointToBox(GlobalVar.ps().boxToPoint(new BoxPoint(i,j)));
-          paintBox(p.getX(),p.getY(),c);
+          paintBox(((int)p.getX()),(int)(p.getY()),c);
         }
       }
     }
